rename getter => finder
