+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/command_handlers.yml
@@ -1,6 +1,6 @@
 services:
   Akeneo\Pim\Enrichment\Component\Product\Commands\GroupProductsHandler:
     arguments:
-      - '@akeneo.pim.enrichment.group.query.get_group_product_identifiers'
+      - '@akeneo.pim.enrichment.product.query.find_product_identifiers'
       - '@doctrine.orm.entity_manager'
       - '@pim_catalog.saver.product'
