+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/normalizers.yml
@@ -61,7 +61,7 @@ services:
             - '@pim_structure_version.provider.structure_version.group'
             - '@pim_versioning.manager.version'
             - '@pim_enrich.normalizer.version'
-            - '@akeneo.pim.enrichment.group.query.get_group_product_identifiers'
+            - '@akeneo.pim.enrichment.product.query.find_product_identifiers'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
