+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/Product/FindProductIdentifiers.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Group;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Product;
 
-use Akeneo\Pim\Enrichment\Component\Product\Query\GetGroupProductIdentifiers;
+use Akeneo\Pim\Enrichment\Component\Product\Query\FindProductIdentifiersInterface;
 use Doctrine\DBAL\Connection;
 
 /**
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class SqlGetGroupProductIdentifiers implements GetGroupProductIdentifiers
+class FindProductIdentifiers implements FindProductIdentifiersInterface
 {
     /** @var Connection */
     private $dbConnection;
