+++ src/Akeneo/Pim/Enrichment/Component/Product/Query/FindProductIdentifiersInterface.php
@@ -8,7 +8,7 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Query;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface GetGroupProductIdentifiers
+interface FindProductIdentifiersInterface
 {
     /**
      * @param int $groupId
