+++ tests/back/Pim/Enrichment/Specification/Component/Product/Commands/GroupProductsHandlerSpec.php
@@ -6,14 +6,14 @@ use Akeneo\Pim\Enrichment\Component\Product\Commands\GroupProductsCommand;
 use Akeneo\Pim\Enrichment\Component\Product\Commands\GroupProductsHandler;
 use Akeneo\Pim\Enrichment\Component\Product\Model\Group;
 use Akeneo\Pim\Enrichment\Component\Product\Model\Product;
-use Akeneo\Pim\Enrichment\Component\Product\Query\GetGroupProductIdentifiers;
+use Akeneo\Pim\Enrichment\Component\Product\Query\FindProductIdentifiersInterface;
 use Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface;
 use Doctrine\ORM\EntityManager;
 use PhpSpec\ObjectBehavior;
 
 class GroupProductsHandlerSpec extends ObjectBehavior
 {
-    function let(GetGroupProductIdentifiers $getGroupProductIdentifiers, EntityManager $entityManager, BulkSaverInterface $productSaver) {
+    function let(FindProductIdentifiersInterface $getGroupProductIdentifiers, EntityManager $entityManager, BulkSaverInterface $productSaver) {
         $this->beConstructedWith($getGroupProductIdentifiers,$entityManager,$productSaver);
     }
     function it_is_initializable()
@@ -21,7 +21,7 @@ class GroupProductsHandlerSpec extends ObjectBehavior
         $this->shouldHaveType(GroupProductsHandler::class);
     }
 
-    function it_can_propagate_product_added_group(GetGroupProductIdentifiers $getGroupProductIdentifiers,  EntityManager $entityManager, BulkSaverInterface $productSaver) {
+    function it_can_propagate_product_added_group(FindProductIdentifiersInterface $getGroupProductIdentifiers, EntityManager $entityManager, BulkSaverInterface $productSaver) {
         $groupProductsCommand = new GroupProductsCommand(1,["productId1", "productId2"] );
         $group = new Group();
         $product2 = new Product();
@@ -33,7 +33,7 @@ class GroupProductsHandlerSpec extends ObjectBehavior
         $productSaver->saveAll([$product2])->shouldHaveBeenCalled();
 
     }
-    function it_can_propagate_product_removed_group(GetGroupProductIdentifiers $getGroupProductIdentifiers,  EntityManager $entityManager, BulkSaverInterface $productSaver) {
+    function it_can_propagate_product_removed_group(FindProductIdentifiersInterface $getGroupProductIdentifiers, EntityManager $entityManager, BulkSaverInterface $productSaver) {
         $groupProductsCommand = new GroupProductsCommand(1,["productId2"] );
         $group = new Group();
         $product2 = new Product();
